+++ src/main/java/dev/krona/urbex/worldgen/CityGenerator.java
@@ -1986,7 +1986,7 @@ public class CityGenerator {
             // how many spawners this chunk happened to place before it.
             RandomSource spawnerRandom = Rng.atPos(provider.getSeed(), pos.getX(), pos.getY(), pos.getZ(), Rng.Purpose.SPAWNERS);
             Identifier randomValue = getRandomSpawnerMob(world.getLevel(), spawnerRandom, provider, info,
-                    new BuildingInfo.ConditionTodo(mobid, ConditionContext.legacyMatchKey(part.getId()), info), pos);
+                    new BuildingInfo.ConditionTodo(mobid, part.getName(), info), pos);
             CompoundTag sd = new CompoundTag();
             sd.putString("id", randomValue.toString());
             SpawnData data = new SpawnData(sd, Optional.empty(), Optional.empty());
@@ -2009,7 +2009,7 @@ public class CityGenerator {
             if (!inWorld.getBlockState(pos).isAir()) {
                 inWorld.setBlock(pos, block, Block.UPDATE_CLIENTS);
                 generateLoot(info, inWorld, pos,
-                        new BuildingInfo.ConditionTodo(marker.loot(), ConditionContext.legacyMatchKey(part.getId()), info));
+                        new BuildingInfo.ConditionTodo(marker.loot(), part.getName(), info));
             }
         });
     }
@@ -2078,7 +2078,7 @@ public class CityGenerator {
         Condition cnd = AssetRegistries.CONDITIONS.getOrThrow(world, condition);
         int level = (pos.getY() - diminfo.getProfile().GROUNDLEVEL) / FLOORHEIGHT;
         int floor = (pos.getY() - info.getCityGroundLevel()) / FLOORHEIGHT;
-        String belowFloor = "<none>";
+        String belowFloor = ConditionContext.NO_PART;
         ConditionContext conditionContext = new ConditionContext(level, floor, info.cellars, info.getNumFloors(),
                 todo.getPart(), belowFloor, todo.getBuilding(), info.coord) {
             @Override
@@ -2114,7 +2114,7 @@ public class CityGenerator {
                 int level = (pos.getY() - diminfo.getProfile().GROUNDLEVEL) / FLOORHEIGHT;
                 int floor = (pos.getY() - info.getCityGroundLevel()) / FLOORHEIGHT;
                 ConditionContext conditionContext = new ConditionContext(level, floor, info.cellars, info.getNumFloors(),
-                        todo.getPart(), "<none>", todo.getBuilding(), info.coord) {
+                        todo.getPart(), ConditionContext.NO_PART, todo.getBuilding(), info.coord) {
                     @Override
                     public Identifier getBiome() {
                         return world.getBiome(pos).unwrap().map(ResourceKey::identifier, biome -> world.registryAccess().lookupOrThrow(Registries.BIOME).getKey(biome));